chore(rig-bedrock): move inline test modules to sibling files
